Add render thumbnails
